Quick Maths done!
Need to figure out the CSS.
Turns out that using a setTimeout that calls itself is stupid, and I for some reason learned to do it early on when starting JavaScript. Did you know there is a built in setInterval method? I ded, but for some reason did not think to use it...