Started work on Personal Space
Set it up with custom hooks for sharing values, and slowly adding variables and functionality.

Also noticed a bug in HurdleRun where two hurdles would sometimes spawn adjacent to eachother, making surviving impossible. Said that hurdle cannot spawn if the first cell is a hurdle, and haven't noticed it since.