Removed display of local points
They were nice when testing out to see if it added up right, but now it's all about the total score.
That, and I couldn't be bothered exporting the points to update elsewhere, and there was no space for them inside PS.

TODO: remove warnings and sort with proper file names and folders!
Ooooooor I knew from the start this would be messy, I'm happy it works, now that I know more, I can check some naming schemes and sorting conventions for the next project.. That sounds better.